refactor(api): Restructure BaseApiClient and BaseBuilder for improved organization and encapsulation
